# Fix clinic activity logs performance by adding index - created-by-agentic