--- a/src/main/resources/db/migration/V2024_01_25_01__add_clinic_activity_logs_index.sql
+++ b/src/main/resources/db/migration/V2024_01_25_01__add_clinic_activity_logs_index.sql
@@ -1,0 +1,2 @@
+-- Add index to improve performance of numeric_value queries
+CREATE INDEX IF NOT EXISTS idx_clinic_activity_logs_numeric_value ON clinic_activity_logs(numeric_value);