--- a/src/main/resources/db/migration/V2__add_clinic_activity_logs_index.sql
+++ b/src/main/resources/db/migration/V2__add_clinic_activity_logs_index.sql
@@ -1,0 +1,2 @@
+-- Add index on numeric_value column to improve query performance
+CREATE INDEX IF NOT EXISTS idx_clinic_activity_logs_numeric_value ON clinic_activity_logs(numeric_value);