--- a/src/main/resources/db/migration/V3__Add_clinic_activity_logs_numeric_value_index.sql
+++ b/src/main/resources/db/migration/V3__Add_clinic_activity_logs_numeric_value_index.sql
@@ -1,0 +1,1 @@
+CREATE INDEX IF NOT EXISTS idx_clinic_activity_logs_numeric_value ON clinic_activity_logs(numeric_value);