This PR addresses a critical performance issue in the Activity Logs query affecting the /api/clinic-activity/query-logs endpoint.

Problem:
- Query performance degradation (P50: 2.38s, P95: 7.8s)
- Full table scan on clinic_activity_logs when filtering by numeric_value
- No index exists on the numeric_value column

Solution:
- Added a B-tree index on the numeric_value column
- Performance improvement verified:
  - Before: Sequential scan with execution time ~4.9s
  - After: Index scan with execution time ~8.4ms (>500x improvement)

Migration:
- Added V4__Add_clinic_activity_logs_numeric_value_index.sql to create the index
- Uses Flyway migration to ensure consistent deployment

Testing:
- Verified query execution plan shows index usage
- Confirmed significant performance improvement in query execution time